refactor: update themeType filename
