+++ apps/desktop/src/services/electron-main-messaging.service.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 import { app, dialog, ipcMain, Menu, MenuItem, nativeTheme, session, Notification } from "electron";
 
 import { MessagingService } from "@bitwarden/common/abstractions/messaging.service";
-import { ThemeType } from "@bitwarden/common/enums/themeType";
+import { ThemeType } from "@bitwarden/common/enums/theme-type";
 
 import { WindowMain } from "../main/window.main";
 import { RendererMenuItem } from "../utils";
