+++ apps/web/src/app/core/state/global-state.ts
@@ -1,4 +1,4 @@
-import { ThemeType } from "@bitwarden/common/enums/themeType";
+import { ThemeType } from "@bitwarden/common/enums/theme-type";
 import { GlobalState as BaseGlobalState } from "@bitwarden/common/models/domain/global-state";
 
 export class GlobalState extends BaseGlobalState {
