+++ libs/angular/src/services/theming/theme.ts
@@ -1,4 +1,4 @@
-import { ThemeType } from "@bitwarden/common/enums/themeType";
+import { ThemeType } from "@bitwarden/common/enums/theme-type";
 
 export interface Theme {
   configuredTheme: ThemeType;
