+++ libs/angular/src/services/theming/themeBuilder.ts
@@ -1,4 +1,4 @@
-import { ThemeType } from "@bitwarden/common/enums/themeType";
+import { ThemeType } from "@bitwarden/common/enums/theme-type";
 
 import { Theme } from "./theme";
