+++ libs/common/src/abstractions/state.service.ts
@@ -10,7 +10,7 @@ import { CollectionView } from "../admin-console/models/view/collection.view";
 import { EnvironmentUrls } from "../auth/models/domain/environment-urls";
 import { KdfConfig } from "../auth/models/domain/kdf-config";
 import { KdfType } from "../enums/kdf-type";
-import { ThemeType } from "../enums/themeType";
+import { ThemeType } from "../enums/theme-type";
 import { UriMatchType } from "../enums/uriMatchType";
 import { EventData } from "../models/data/event.data";
 import { ServerConfigData } from "../models/data/server-config.data";
