+++ libs/common/src/services/stateMigration.service.ts
@@ -8,7 +8,7 @@ import { TokenService } from "../auth/services/token.service";
 import { HtmlStorageLocation } from "../enums/html-storage-location";
 import { KdfType } from "../enums/kdf-type";
 import { StateVersion } from "../enums/state-version";
-import { ThemeType } from "../enums/themeType";
+import { ThemeType } from "../enums/theme-type";
 import { StateFactory } from "../factories/stateFactory";
 import { EventData } from "../models/data/event.data";
 import {
